fix: users.service

+++ src/users/users.service.ts
@@ -2,6 +2,8 @@ import {
   Injectable,
   NotFoundException,
   BadRequestException,
+  Logger,
+  InternalServerErrorException,
 } from '@nestjs/common';
 
 import { Repository } from 'typeorm';
@@ -14,6 +16,8 @@ import { UpdateUserDto } from './dto/update-user.dto';
 
 @Injectable()
 export class UsersService {
+  private readonly logger = new Logger('UserService');
+
   constructor(
     @InjectRepository(User)
     private readonly userRepository: Repository<User>,
@@ -21,39 +25,65 @@ export class UsersService {
 
   async create(createUserDto: CreateUserDto) {
     const { email } = createUserDto;
-    const existUsers = await this.findByEmail(email);
+    
+    try {
+      const existUsers = await this.findByEmail(email);
 
-    if (existUsers) {
-      throw new BadRequestException(`user with email ${email} exist`);
+      if (existUsers) {
+        throw new BadRequestException(`user with email ${email} exist`);
+      }
+
+      return await this.userRepository.save(createUserDto);
+    } catch (error) {
+      this.logger.error(error.message);
+
+      if (error instanceof BadRequestException) {
+        throw error;
+      }
+
+      throw new InternalServerErrorException('create user error');
     }
-
-    return await this.userRepository.save(createUserDto);
   }
 
   async findAll() {
-    return await this.userRepository.find({
-      // relations: ['roles'],
-      select: {
-        id: true,
-        name: true,
-        about: true,
-        avatar: true,
-      },
-    });
+    try {
+      return await this.userRepository.find({
+        // relations: ['roles'],
+        select: {
+          id: true,
+          name: true,
+          about: true,
+          avatar: true,
+        },
+      });
+    } catch (error) {
+      this.logger.error(error.message);
+      throw new InternalServerErrorException(error.message);
+    }
   }
 
   async findOne(id: number) {
-    const user = await this.userRepository.findOne({
-      where: { id },
-      relations: ['roles'],
-    });
+    try {
+      if (!Number.isInteger(id) || id <= 0) {
+        throw new BadRequestException('Invalid user ID');
+      }
 
-    if (!user) {
-      throw new NotFoundException(`user with id ${id} not found`);
+      const user = await this.userRepository.findOne({
+        where: { id },
+        relations: ['roles'],
+      });
+
+      if (!user) {
+        throw new NotFoundException(`user with id ${id} not found`);
+      }
+
+      return user;
+    } catch (error) {
+      this.logger.error('Update user error');
+      throw error;
     }
-
-    return user;
   }
+
   async findByAvatar(fileName: string) {
     const user = await this.userRepository.findOne({
       where: { avatar: fileName },
@@ -78,25 +108,73 @@ export class UsersService {
   }
 
   async update(updateUserDto: UpdateUserDto) {
-    await this.userRepository.update(updateUserDto.id, {
-      name: updateUserDto.name,
-      about: updateUserDto.about,
-    });
+    try {
+      const user = await this.userRepository.findOne({
+        where: { id: updateUserDto.id },
+      });
 
-    const user = await this.userRepository.findOne({
-      where: { id: updateUserDto.id },
-    });
+      if (!user) {
+        throw new Error('User not found');
+      }
 
-    return user;
+      this.logger.log(`Update user ${updateUserDto.id}`);
+
+      await this.userRepository.update(updateUserDto.id, {
+        name: updateUserDto.name,
+        about: updateUserDto.about,
+      });
+
+      return { ...user,
+        name: updateUserDto.name,
+        about: updateUserDto.about,
+      };
+    } catch (error) {
+      this.logger.error('Update user error');
+      throw error;
+    }
   }
 
   async updateAvatar(updateUserDto: { id: number; avatar: string }) {
-    return this.userRepository.update(updateUserDto.id, {
-      avatar: updateUserDto.avatar,
-    });
+    try {
+      const user = await this.userRepository.findOne({
+        where: { id: updateUserDto.id },
+      });
+
+      if (!user) {
+        throw new Error('User not found');
+      }
+
+      this.logger.log(`Update user avatar ${updateUserDto.id}`);
+
+      return this.userRepository.update(updateUserDto.id, {
+        avatar: updateUserDto.avatar,
+      });
+    } catch (error) {
+      this.logger.error('Update avatar error');
+      throw error;
+    }
   }
 
-  remove(id: number) {
-    return this.userRepository.delete(id);
+  async remove(id: number) {
+    if (!Number.isInteger(id) || id <= 0) {
+      throw new BadRequestException('Invalid user ID');
+    }
+
+    try {
+      const user = await this.userRepository.findOne({
+        where: { id },
+      });
+
+      if (!user) {
+        throw new Error('User not found');
+      }
+
+      this.logger.log(`Удаление пользователя с ID: ${id}`);
+      
+      return this.userRepository.delete(id);
+    } catch (error) {
+      this.logger.error(`Delete user with ID: ${id}`);
+      throw error;
+    }
   }
 }
